fix:Fix the issue of not generating irrelevant data
